fix: switch from `$base..$old $base..$new` to `$base $old $new`
if we went far enough with $old, then what happen was
$base became equal to $old
and giving a range of `$base..$old` would error out range-diff,
since that's not a range because it's just 1 commit.

using the new approach avoids this error.

one could argue that a range of the same commit A..A is still a range,
and could implement a fix into core git..